+++ Kokoro/API/StatusNextTurn.cs
@@ -1,0 +1,12 @@
+﻿using CobaltCoreModding.Definitions.ExternalItems;
+
+namespace Shockah.Kokoro;
+
+public partial interface IKokoroApi
+{
+	ExternalStatus TempShieldNextTurnStatus { get; }
+	Status TempShieldNextTurnVanillaStatus { get; }
+
+	ExternalStatus ShieldNextTurnStatus { get; }
+	Status ShieldNextTurnVanillaStatus { get; }
+}